subscribe to currency price

+++ coingecko/models.go
@@ -4,6 +4,6 @@ type Price struct {
 	Usd           float64 `json:"usd"`
 	UsdMarketCap  float64 `json:"usd_market_cap,omitempty"`
 	Usd24HChange  float64 `json:"usd_24h_change,omitempty"`
-	LastUpdatedAt int     `json:"last_updated_at,omitempty"`
+	LastUpdatedAt int64   `json:"last_updated_at,omitempty"`
 	Currency      string  `json:"currency,omitempty"`
 }
